Move key handlers into their own file, and add to oval and rect tools
